requirements.txt
tqdm
urllib3==1.25.3
feedparser==6.0.8
pyyaml
pysftp